restgen/app/requirements.txt
click==6.7
Flask==0.12.2
Flask-Restless==0.17.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mimerender==0.6.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.2.2
Werkzeug==0.14.1
